tests: move tests of `Ui` from `tests/` into `ui` module
The tests only test the `Ui` type and don't need anything from
`testutils`, so it seems they should be treated as unit tests.
